Addressing pr feedback
