+++ ingestion_tools/scripts/data_validation/shared/helper/mdoc_helper.py
@@ -17,7 +17,10 @@ class MdocTestHelper:
 
     @allure.title("Mdoc: number of mdoc sections equal number of frames files.")
     def test_mdoc_frames(self, mdoc_data: pd.DataFrame, frames_files: list[str]):
-        assert len(mdoc_data) == len(frames_files)
+        mdoc_len = len(mdoc_data)
+        frames_len = len(frames_files)
+        assert mdoc_len == frames_len, f"Number of mdoc sections {mdoc_len} mismatches number of frames: {frames_len}"
+
 
     @allure.title("Mdoc: Every mdoc SubFramePath filename matches a frames file (one-to-one).")
     def test_mdoc_frame_paths(
